# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Associate Onboarding Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Cohort {
  id: number;
  name: string;
  /** Calendar date YYYY-MM-DD */
  startDate: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface CohortInput {
  /** @minLength 1 */
  name: string;
  /** Calendar date YYYY-MM-DD */
  startDate: string;
  description?: string;
}

export interface CohortUpdate {
  /** @minLength 1 */
  name?: string;
  startDate?: string;
  description?: string;
}

export type SupervisorRole = typeof SupervisorRole[keyof typeof SupervisorRole];


export const SupervisorRole = {
  director: 'director',
  manager: 'manager',
  team_lead: 'team_lead',
  senior_mentor: 'senior_mentor',
} as const;

export interface Supervisor {
  id: number;
  name: string;
  email: string;
  role: SupervisorRole;
  /** @nullable */
  title?: string | null;
  createdAt: string;
}

export type SupervisorInputRole = typeof SupervisorInputRole[keyof typeof SupervisorInputRole];


export const SupervisorInputRole = {
  director: 'director',
  manager: 'manager',
  team_lead: 'team_lead',
  senior_mentor: 'senior_mentor',
} as const;

export interface SupervisorInput {
  /** @minLength 1 */
  name: string;
  email: string;
  role: SupervisorInputRole;
  title?: string;
}

export type SupervisorUpdateRole = typeof SupervisorUpdateRole[keyof typeof SupervisorUpdateRole];


export const SupervisorUpdateRole = {
  director: 'director',
  manager: 'manager',
  team_lead: 'team_lead',
  senior_mentor: 'senior_mentor',
} as const;

export interface SupervisorUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  role?: SupervisorUpdateRole;
  title?: string;
}

export interface Associate {
  id: number;
  name: string;
  email: string;
  cohortId: number;
  /** @nullable */
  directorId?: number | null;
  /** @nullable */
  managerId?: number | null;
  /** @nullable */
  teamLeadId?: number | null;
  /** @nullable */
  seniorMentorId?: number | null;
  /** @nullable */
  position?: string | null;
  /** @nullable */
  department?: string | null;
  createdAt: string;
  cohort?: Cohort;
  director?: Supervisor;
  manager?: Supervisor;
  teamLead?: Supervisor;
  seniorMentor?: Supervisor;
}

export interface AssociateInput {
  /** @minLength 1 */
  name: string;
  email: string;
  cohortId: number;
  directorId?: number;
  managerId?: number;
  teamLeadId?: number;
  seniorMentorId?: number;
  position?: string;
  department?: string;
}

export interface AssociateUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  cohortId?: number;
  /** @nullable */
  directorId?: number | null;
  /** @nullable */
  managerId?: number | null;
  /** @nullable */
  teamLeadId?: number | null;
  /** @nullable */
  seniorMentorId?: number | null;
  position?: string;
  department?: string;
}

export type PhaseProgressPhase = typeof PhaseProgressPhase[keyof typeof PhaseProgressPhase];


export const PhaseProgressPhase = {
  pre_start: 'pre_start',
  first_day: 'first_day',
  week_1: 'week_1',
  week_2: 'week_2',
  week_3: 'week_3',
  week_4: 'week_4',
} as const;

export type CheckinSupervisorRole = typeof CheckinSupervisorRole[keyof typeof CheckinSupervisorRole];


export const CheckinSupervisorRole = {
  director: 'director',
  manager: 'manager',
  team_lead: 'team_lead',
  senior_mentor: 'senior_mentor',
} as const;

export type CheckinPhase = typeof CheckinPhase[keyof typeof CheckinPhase];


export const CheckinPhase = {
  pre_start: 'pre_start',
  first_day: 'first_day',
  week_1: 'week_1',
  week_2: 'week_2',
  week_3: 'week_3',
  week_4: 'week_4',
} as const;

export type CheckinStatus = typeof CheckinStatus[keyof typeof CheckinStatus];


export const CheckinStatus = {
  pending: 'pending',
  completed: 'completed',
} as const;

export interface Checkin {
  id: number;
  associateId: number;
  supervisorId: number;
  supervisorRole: CheckinSupervisorRole;
  phase: CheckinPhase;
  status: CheckinStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
  supervisor?: Supervisor;
  associate?: Associate;
}

export interface PhaseProgress {
  phase: PhaseProgressPhase;
  completedCount?: number;
  totalCount?: number;
  checkins: Checkin[];
}

export interface AssociateProgress {
  associateId: number;
  totalCheckins?: number;
  completedCheckins?: number;
  overallPercent?: number;
  phases: PhaseProgress[];
}

export type CheckinInputSupervisorRole = typeof CheckinInputSupervisorRole[keyof typeof CheckinInputSupervisorRole];


export const CheckinInputSupervisorRole = {
  director: 'director',
  manager: 'manager',
  team_lead: 'team_lead',
  senior_mentor: 'senior_mentor',
} as const;

export type CheckinInputPhase = typeof CheckinInputPhase[keyof typeof CheckinInputPhase];


export const CheckinInputPhase = {
  pre_start: 'pre_start',
  first_day: 'first_day',
  week_1: 'week_1',
  week_2: 'week_2',
  week_3: 'week_3',
  week_4: 'week_4',
} as const;

export interface CheckinInput {
  associateId: number;
  supervisorId: number;
  supervisorRole: CheckinInputSupervisorRole;
  phase: CheckinInputPhase;
  notes?: string;
}

export type CheckinUpdateStatus = typeof CheckinUpdateStatus[keyof typeof CheckinUpdateStatus];


export const CheckinUpdateStatus = {
  pending: 'pending',
  completed: 'completed',
} as const;

export interface CheckinUpdate {
  notes?: string;
  status?: CheckinUpdateStatus;
}

export interface CheckinCompletion {
  notes: string;
}

export interface DashboardSummary {
  totalCohorts: number;
  totalAssociates: number;
  totalSupervisors: number;
  totalCheckins: number;
  completedCheckins: number;
  pendingCheckins: number;
  /** Percentage 0-100 */
  completionRate: number;
  recentlyCompleted?: Checkin[];
}

export interface CohortDashboard {
  cohort: Cohort;
  totalAssociates?: number;
  associates: AssociateProgress[];
  totalCheckins: number;
  completedCheckins: number;
  completionRate: number;
}

export type PendingCheckinSupervisorRole = typeof PendingCheckinSupervisorRole[keyof typeof PendingCheckinSupervisorRole];


export const PendingCheckinSupervisorRole = {
  director: 'director',
  manager: 'manager',
  team_lead: 'team_lead',
  senior_mentor: 'senior_mentor',
} as const;

export type PendingCheckinPhase = typeof PendingCheckinPhase[keyof typeof PendingCheckinPhase];


export const PendingCheckinPhase = {
  pre_start: 'pre_start',
  first_day: 'first_day',
  week_1: 'week_1',
  week_2: 'week_2',
  week_3: 'week_3',
  week_4: 'week_4',
} as const;

export interface PendingCheckin {
  id: number;
  associateId: number;
  associateName: string;
  cohortName?: string;
  supervisorId: number;
  supervisorName: string;
  supervisorRole: PendingCheckinSupervisorRole;
  phase: PendingCheckinPhase;
  createdAt: string;
}

export type ListAssociatesParams = {
cohortId?: number;
};

export type ListCheckinsParams = {
associateId?: number;
supervisorId?: number;
phase?: ListCheckinsPhase;
};

export type ListCheckinsPhase = typeof ListCheckinsPhase[keyof typeof ListCheckinsPhase];


export const ListCheckinsPhase = {
  pre_start: 'pre_start',
  first_day: 'first_day',
  week_1: 'week_1',
  week_2: 'week_2',
  week_3: 'week_3',
  week_4: 'week_4',
} as const;