# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Associate Onboarding Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /supervisors:
    get:
      operationId: listSupervisors
      tags: [supervisors]
      summary: List all supervisors
      responses:
        "200":
          description: List of supervisors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supervisor"
    post:
      operationId: createSupervisor
      tags: [supervisors]
      summary: Create a new supervisor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupervisorInput"
      responses:
        "201":
          description: Supervisor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supervisor"
        "400":
          description: Invalid input
  /supervisors/{id}:
    get:
      operationId: getSupervisor
      tags: [supervisors]
      summary: Get a supervisor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Supervisor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supervisor"
        "404":
          description: Not found
    patch:
      operationId: updateSupervisor
      tags: [supervisors]
      summary: Update a supervisor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupervisorUpdate"
      responses:
        "200":
          description: Updated supervisor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supervisor"
        "404":
          description: Not found
    delete:
      operationId: deleteSupervisor
      tags: [supervisors]
      summary: Delete a supervisor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ─── Associates ───────────────────────────────────────────────────────────
...
  /dashboard/cohort/{cohortId}:
    get:
      operationId: getCohortDashboard
      tags: [dashboard]
      summary: Get per-cohort progress breakdown
      parameters:
        - name: cohortId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Cohort dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CohortDashboard"
        "404":
          description: Not found
...
components:
  schemas:
...
    Cohort:
      type: object
      required: [id, name, startDate, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        startDate:
          type: string
          description: Calendar date YYYY-MM-DD
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Supervisor:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [director, manager, team_lead, senior_mentor]
        title:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    SupervisorInput:
      type: object
      required: [name, email, role]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        role:
          type: string
          enum: [director, manager, team_lead, senior_mentor]
        title:
          type: string
    SupervisorUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        role:
          type: string
          enum: [director, manager, team_lead, senior_mentor]
        title:
          type: string

    # ─── Associate ──────────────────────────────────────────────────────────
    Associate:
      type: object
      required: [id, name, email, cohortId, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        cohortId:
          type: number
        directorId:
          type: ["number", "null"]
        managerId:
          type: ["number", "null"]
        teamLeadId:
          type: ["number", "null"]
        seniorMentorId:
          type: ["number", "null"]
        position:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        # Hydrated relationships (optional, returned with detail queries)
        cohort:
          $ref: "#/components/schemas/Cohort"
        director:
          $ref: "#/components/schemas/Supervisor"
        manager:
          $ref: "#/components/schemas/Supervisor"
        teamLead:
          $ref: "#/components/schemas/Supervisor"
        seniorMentor:
          $ref: "#/components/schemas/Supervisor"
...
    AssociateProgress:
      type: object
      required: [associateId, phases]
      properties:
        associateId:
          type: number
        totalCheckins:
          type: number
        completedCheckins:
          type: number
        overallPercent:
          type: number
        phases:
          type: array
          items:
            $ref: "#/components/schemas/PhaseProgress"
    PhaseProgress:
      type: object
      required: [phase, checkins]
      properties:
        phase:
          type: string
          enum: [pre_start, first_day, week_1, week_2, week_3, week_4]
        completedCount:
          type: number
        totalCount:
          type: number
        checkins:
          type: array
          items:
            $ref: "#/components/schemas/Checkin"

    # ─── Checkin ────────────────────────────────────────────────────────────
    Checkin:
      type: object
      required: [id, associateId, supervisorId, supervisorRole, phase, status, createdAt]
      properties:
        id:
          type: number
        associateId:
          type: number
        supervisorId:
          type: number
        supervisorRole:
          type: string
          enum: [director, manager, team_lead, senior_mentor]
        phase:
          type: string
          enum: [pre_start, first_day, week_1, week_2, week_3, week_4]
        status:
          type: string
          enum: [pending, completed]
        notes:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        # Hydrated
        supervisor:
          $ref: "#/components/schemas/Supervisor"
        associate:
          $ref: "#/components/schemas/Associate"
...
    CohortDashboard:
      type: object
      required: [cohort, associates, totalCheckins, completedCheckins, completionRate]
      properties:
        cohort:
          $ref: "#/components/schemas/Cohort"
        totalAssociates:
          type: number
        associates:
          type: array
          items:
            $ref: "#/components/schemas/AssociateProgress"
        totalCheckins:
          type: number
        completedCheckins:
          type: number
        completionRate:
          type: number
...