# lib/api-zod/src/generated/types/supervisorUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Associate Onboarding Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SupervisorUpdateRole } from './supervisorUpdateRole';

export interface SupervisorUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  role?: SupervisorUpdateRole;
  title?: string;
}